Added Caching
Here, I added a caching feature. It overrides the existing translated file, if the key already exist in the previously translated file, it doesn't make the request to the Google Translate API, instead it reuses the value. The only condition is that you keep your files in the tmp folder on the root level of the project and when you update your original JSON File it will check bot the updated JSON and updates the older translated versions with new keys and values.